Q4 form scoring project/backend/auth/requirements.txt
# Authentication API - Cloud Function Dependencies
# Created: November 5, 2025

# Google Cloud Platform
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0

# Authentication & Security
pyjwt>=2.8.0
bcrypt>=4.0.0
cryptography>=41.0.0

# Configuration
python-dotenv>=1.0.0
